Show attachment link in body area
