+++ components/article/ArticleAttachments.js
@@ -1,37 +1,6 @@
 import React from 'react'
+import formatFileLinkText from '../../helpers/attachments'
 import { fonts, fontSizes, measurements } from '../../theme/variables'
-
-const formatFileSize = (size) => {
-  if (!size) { return '' }
-  let filesize = Math.round(size / 1000)
-  if (filesize < 1000) {
-    return `${filesize} kB`
-  }
-  filesize = (filesize / 1000).toFixed(2)
-  return `${filesize} MB`
-}
-
-const formatMimeType = (mimeType) => {
-  if (!mimeType) { return '' }
-  return mimeType.split('/')[1].toUpperCase()
-}
-
-const formatFileLinkText = (file) => {
-  const name = file.description || file.filename || file.url
-  return (
-    <span>
-      <strong>{name}</strong> ({formatMimeType(file.mimetype)} <span className='divider'>|</span> {formatFileSize(file.filesize)})
-      <style jsx>{`
-        span {
-          font-weight: normal;
-        }
-        .divider {
-          padding: 0 ${measurements.baseUnit / 2}em;
-        }
-      `}</style>
-    </span>
-  )
-}
 
 class ArticleAttachments extends React.Component {
   render () {
